implemeted graph class, dijkstra algorithm for shorted paths and metroGraph class for finding shortest paths in metro

+++ app/app/src/main/java/app/gatherround/metro/Link.kt
@@ -9,5 +9,5 @@ data class Link(
     val toStationId: Int,
     val weightTime: Int,
     val weightTransfer: Int,
-    val transferId: Int? = null
+    val transferId: Long? = null
 )
